Enhsp improvement

+++ setup.py
@@ -1,5 +1,8 @@
 #!/usr/bin/env python3
+from ast import Raise
 import subprocess
+import warnings
+import sys
 
 from setuptools import setup
 from setuptools.command.build_py import build_py
@@ -14,6 +17,7 @@ ENHSP_PUBLIC = 'ENHSP-Public'
 COMPILE_CMD = './compile'
 ENHSP_TAG = 'enhsp20-0.9.4'
 ENHSP_REPO = 'https://gitlab.com/enricos83/ENHSP-Public'
+JDK_REQUIRE = 17
 
 long_description = \
     """============================================================
@@ -21,8 +25,17 @@ long_description = \
  ============================================================
 """
 
+def check_version_jdk():
+    result = subprocess.check_output(['java', '--version']).decode('utf-8') 
+    versione = result.split(" ")[1].split(".")[0]
+    if int(versione) >= JDK_REQUIRE:
+        return True
+    else:
+        return False
 
 def install_ENHSP():
+    if not check_version_jdk():
+        raise ValueError(f'ENHSP require jdk version >= {JDK_REQUIRE}')
     subprocess.run(['git', 'clone', '-b', ENHSP_TAG, ENHSP_REPO])
     shutil.rmtree(ENHSP_dst,ignore_errors=True)
     shutil.move(ENHSP_PUBLIC, ENHSP_dst)
